feat: Featured component
